Fix reward being returned early after the first episode
With commit 22b47fb67b95b08039fa335bed6053d97c7052f9, `train` was now
returning the mean reward but for the A2C and DQN agents the reward is
being returned prematurly. Only one episode runs and with this change
the mean reward is returned after all the steps and episodes are
completed.